Implement the HTML and initial styling.
I created a div for the description text, another to hold the buttons
and another to hold the player's move + computer's move and a versus
image.
The initial reset was created (It might need some reavaliation) and the
buttons are responsive. But there is still room for improvements and
make the image and text inside the buttons responsive.